python-service/requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==4.23.1